==> scripts/00-simulate_data.ipynb <==
#### Preamble ####
# Purpose: Simulates some data for BTC/USDT
# Author: Jiazhou(Justin) Bi
# Date: Nov 15, 2024
# Contact: justin.bi@mail.utoronto.ca
# License: None
# Pre-requisites: see requirements.txt
# Any other information needed? None

In the following section, I will simulate 1000 observations for my desired dataset based on the sketch in the "other" folder, as well as logics of the data columns. The simulated data with be saved to "../data/00-simulated_data/simulated_data.parquet" and tested in the "/./01-test_simulated_data.ipynb/" in the same folder of this notebook.

import pandas as pd
import numpy as np


# Creating an empty DataFrme for the simulation
simulated_df = pd.DataFrame(columns=['Time','Open-Diff','Low-Diff','High-Diff','Close-Diff','Volume','D_t-1','D_t+1'])
simulated_df

# Creating the timestamps for the dataset.
# 1-hour interval dataset in this case.
full_time_range = pd.date_range(start=pd.Timestamp("2020-01-01 16:00:00"), end=pd.Timestamp("2022-01-01 16:00:00"), freq='1h')

# Feed the time range to the empty DataFrame and only keep the first 1000 rows
time_values = full_time_range[:1000]
simulated_df['Time'] = time_values
print(simulated_df.shape)
print(simulated_df.head())

Simulating the values for the column "Close-Diff". Here the starting point of its value is near 20, and ends at around 22. Randomness and fluctuations were added to introduce noises to the data.

# Setting parameters up
np.random.seed(210)
num_rows = 1000

# Simulate random fluctuations
close_diff = np.random.normal(loc=0, scale=20, size=num_rows)

# Feed the simulated values to the DataFrame
simulated_df['Close-Diff'] = close_diff

print(simulated_df.head())

Applying the same logic to simulate "Open-Diff", "Low-Diff", "High-Diff", and Volume.

# Setting parameters up
num_rows = 1000
np.random.seed(211)
open_diff = np.random.normal(loc=0, scale=20, size=num_rows)
np.random.seed(212)
low_diff = np.random.normal(loc=0, scale=30, size=num_rows)
np.random.seed(213)
high_diff = np.random.normal(loc=0, scale=30, size=num_rows)
np.random.seed(214)
volume = np.abs(np.random.normal(loc=0, scale=2, size=num_rows))

# Feed the simulated values to the DataFrame
simulated_df['Open-Diff'] = open_diff
simulated_df['Low-Diff'] = low_diff
simulated_df['High-Diff'] = high_diff
simulated_df['Volume'] = volume

print(simulated_df.head())

Lastly, I will calculate the values of D_t-1 and D_t+1. D_t-1 will be -1 if the Close-Diff is negative, and 1 if Close-Diff is positive. For D_t+1, if the next CLose-Diff is positive, it is 1, otherwise it will be 0, as per definition of our dataset.


simulated_df['D_t-1'] = simulated_df['Close-Diff'].apply(lambda x: 1 if x > 0 else -1)
simulated_df['D_t+1'] = (simulated_df['Close-Diff'].shift(-1).apply(lambda x: 1 if pd.notnull(x) and x > 0 else -1 if pd.notnull(x) else None))

simulated_df.head(20)

# Saving the DataFrame into a parquet for testing
simulated_df.to_parquet('../data/00-simulated_data/simulated_data_1h.parquet', index=False)
