==> scripts/02-extract_and_download_data.ipynb <==
#### Preamble ####
# Purpose: Extacts and downloads and saves the historical data of BTC/USDT
# Author: Jiazhou(Justin) Bi
# Date: Nov 15, 2024
# Contact: justin.bi@mail.utoronto.ca
# License: None
# Pre-requisites: see requirements.txt
# Any other information needed? None

# Claim

Currently, this data extraction extract historical data of BTC/USTD from binance on a 1 minute, 1 hour, and 1 day interval, starting 2017-08-17 04:00:00. It seems that binance can handle more calls from API, thus making our data extraction with high granularity (1m compared to 1h or 1d) effortlessly. Currently on Nov 14, 2024, the data retrieval process takes around 22 to 23 mins on my computer. Note that Binance only offers historical data of BTC/USDT started in 2017-08-17. For data with longer historical records, please use other exchanges' data. For more details, please check: https://docs.ccxt.com/#/.

# import necessary packages
import ccxt
import pandas as pd
from datetime import datetime

# Initialize the exchange
exchange = ccxt.binance() #binance, coinbase pro, kraken, bitfinex
symbol = 'BTC/USDT'
timeframe = '1m'      # Timeframe (e.g., '1m', '5m', '15m', '1h', '1d')
since = exchange.parse8601('2017-01-01T00:00:00Z') # Start date in ISO 8601 format
# Binance: Bitcoin historical data is available starting from 2017.
# Coinbase Pro: Data may go back to 2015 for BTC-USD, but it has limitations due to the recent deprecation of Coinbase Pro.
# Bitfinex: Bitcoin historical data goes back to around 2013.
# Kraken: Provides data from as early as 2011 for BTC-USD, though its granularity may vary for older data.

#retrieve data in batches
all_data = []
while since < exchange.milliseconds():
    data = exchange.fetch_ohlcv(symbol, timeframe, since=since, limit=500)
    if not data:
        break
    all_data.extend(data)
    since = data[-1][0] + 1

# this cell takes roughly 22 mins on my PC

df = pd.DataFrame(all_data, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'])
df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
print(df.head())

# saving the 1m timestamp data into parquet
df.to_parquet('../data/01-raw_data/raw_data_1m.parquet', index=False)
# (3804375, 6)

# using the same method to extract data on 1 hour timestamps (intervals)
timeframe = '1h'
since = exchange.parse8601('2017-01-01T00:00:00Z')

#retrieve data in batches
all_data = []
while since < exchange.milliseconds():
    data = exchange.fetch_ohlcv(symbol, timeframe, since=since, limit=500)
    if not data:
        break
    all_data.extend(data)
    since = data[-1][0] + 1

df = pd.DataFrame(all_data, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'])
df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
print(df.head())

# saving the 1h timestamp data into parquet
df.to_parquet('../data/01-raw_data/raw_data_1h.parquet', index=False)
# shape (63423, 6)

print(df.shape)

# using the same method to extract data on 1 day timestamps (intervals)
timeframe = '1d'
since = exchange.parse8601('2017-01-01T00:00:00Z')

#retrieve data in batches
all_data = []
while since < exchange.milliseconds():
    data = exchange.fetch_ohlcv(symbol, timeframe, since=since, limit=500)
    if not data:
        break
    all_data.extend(data)
    since = data[-1][0] + 1

df = pd.DataFrame(all_data, columns=['timestamp', 'open', 'high', 'low', 'close', 'volume'])
df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
print(df.head())

# saving the 1d timestamp data into parquet
df.to_parquet('../data/01-raw_data/raw_data_1d.parquet', index=False)
# shape (2649, 6)
