==> scripts/03-clean_data.ipynb <==
#### Preamble ####
# Purpose: Cleans, engineers, and saves the cleaned data 
# Author: Jiazhou(Justin) Bi
# Date: 15 Nov 2024
# Contact: justin.bi@mail.utoronto.ca
# License: None
# Pre-requisites: see requirements.txt
# Any other information needed? None

# Loading the data

import pandas as pd

#Loading the datasets as DataFrame
df_1m = pd.read_parquet('../data/01-raw_data/raw_data_1m.parquet')
df_1h = pd.read_parquet('../data/01-raw_data/raw_data_1h.parquet')
df_1d = pd.read_parquet('../data/01-raw_data/raw_data_1d.parquet')
# print(df_1m.head())
# print(df_1h.head())
# print(df_1d.head())

# Checking for Missing Values

df_1m.isna().sum()

df_1h.isna().sum()

df_1d.isna().sum()

There are no missing values found in these datasets.

# Data Types

This section examines the datatypes of each column and ensures they are appropriate for the analysis.

df_1m.dtypes

df_1h.dtypes

df_1d.dtypes

# Data Validation

This section is to validate if the datasets are consecutive. That is, for the 1-minute timestamp dataset, no minutes should be skipped. Same logic appleis to the 1-hour and 1-day timestamp datasets.

## Validating if any minute is missing

# creating a full range of minutes for the 1-minute dataset
full_minute_range = pd.date_range(start=df_1m['timestamp'].min(), end=df_1m['timestamp'].max(), freq='1min')
# print(full_minute_range)

# Reindexing the DataFrame
df_1m.set_index('timestamp', inplace=True)
df_1m = df_1m.reindex(full_minute_range)
df_1m.index.name = 'index'
df_1m['index'] = df_1m.index
df_1m.reset_index(drop=True, inplace=True)

# Check for missing values across the entire DataFrame
missing_rows = df_1m[df_1m.isna().any(axis=1)]

if missing_rows.empty:
    print("No missing values.")
else:
    print("Rows with missing values:")
    print(missing_rows)

There are missing minutes. We will first mark the missing rows down in a new column called 'was_miss'. 1 indicates the row was empty and 0 means not. Then the missing values are filled with linear interpolation, except the column "volume", which is replaced with 0.

# creating an indicator to flag the missing rows
df_1m['was_missing'] = df_1m.isna().any(axis=1).astype(int)

# Interpolate missing values
df_1m.interpolate(method='linear', inplace=True)
df_1m.loc[df_1m['was_missing'] == 1, 'volume'] = 0

# Check agian for missing values across the entire DataFrame
missing_rows = df_1m[df_1m.isna().any(axis=1)]

if missing_rows.empty:
    print("No missing values.")
else:
    print("Rows with missing values:")
    print(missing_rows)

## Validating if any hour is missing from the 1-hour dataset

# using the same logic as above
full_hour_range = pd.date_range(start=df_1h['timestamp'].min(), end=df_1h['timestamp'].max(), freq='1h')

# Reindexing the DataFrame
df_1h.set_index('timestamp', inplace=True)
df_1h = df_1h.reindex(full_hour_range)
df_1h.index.name = 'index'
df_1h['index'] = df_1h.index
df_1h.reset_index(drop=True, inplace=True)

# Check for missing values across the entire DataFrame
missing_rows_1h = df_1h[df_1h.isna().any(axis=1)]

if missing_rows_1h.empty:
    print("No missing values.")
else:
    print("Rows with missing values:")
    print(missing_rows_1h)

# applying the same logic as above
df_1h['was_missing'] = df_1h.isna().any(axis=1).astype(int)

# Interpolate missing values
df_1h.interpolate(method='linear', inplace=True)
df_1h.loc[df_1h['was_missing'] == 1, 'volume'] = 0

# Check agian for missing values across the entire DataFrame
missing_rows_1h = df_1h[df_1h.isna().any(axis=1)]

if missing_rows_1h.empty:
    print("No missing values.")
else:
    print("Rows with missing values:")
    print(missing_rows_1h)

## Validating if any day is missing from the 1-day dataset

df_1d.head()

# using the same logic as above
full_day_range = pd.date_range(start=df_1d['timestamp'].min(), end=df_1d['timestamp'].max(), freq='1d')

# Reindexing the DataFrame
df_1d.set_index('timestamp', inplace=True)
df_1d = df_1d.reindex(full_day_range)
df_1d.index.name = 'index'
df_1d['index'] = df_1d.index
df_1d.reset_index(drop=True, inplace=True)

# Check for missing values across the entire DataFrame
missing_rows_1d = df_1d[df_1d.isna().any(axis=1)]

if missing_rows_1d.empty:
    print("No missing values.")
else:
    print("Rows with missing values:")
    print(missing_rows_1d)

For the 1-day dataset, no days were missing.

# Adding A Column For Price Change Direction

This subsection creates two new column for each dataset called direction. For direction_t-1, if the closing price is higher than the previous closing price, it is considered that the price has gone up and thus marked as 1 for appreciation. If the closing price is lower than the previous closing price, it is considered that the price has gone down and hence is marked as -1 for depreciation. If the price remains the same,  it is marked as 0 for no movement. Same for direction_t+1, but this indicates that if the price will go up or down in the next tick(timestamp).

# Calculate direction_t-1 and direction_t+1
df_1m['direction_t-1'] = df_1m['close'].diff().apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))
df_1m['direction_t+1'] = df_1m['close'].shift(-1).sub(df_1m['close']).apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))

# Drop the first and last rows where direction_t-1 or direction_t+1 is NaN
df_1m.dropna(subset=['direction_t-1', 'direction_t+1'], inplace=True)

# Ensure both columns are integers
df_1m['direction_t-1'] = df_1m['direction_t-1'].astype(int)
df_1m['direction_t+1'] = df_1m['direction_t+1'].astype(int)

# Display the DataFrame for verification
df_1m.head(20)

# applying the same logic to the 1-hour and 1-day datasets
df_1h['direction_t-1'] = df_1h['close'].diff().apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))
df_1h['direction_t+1'] = df_1h['close'].shift(-1).sub(df_1h['close']).apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))
df_1h.dropna(subset=['direction_t-1', 'direction_t+1'], inplace=True)
df_1h['direction_t-1'] = df_1h['direction_t-1'].astype(int)
df_1h['direction_t+1'] = df_1h['direction_t+1'].astype(int)

df_1d['direction_t-1'] = df_1d['close'].diff().apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))
df_1d['direction_t+1'] = df_1d['close'].shift(-1).sub(df_1d['close']).apply(lambda x: 1 if x > 0 else (-1 if x < 0 else 0))
df_1d.dropna(subset=['direction_t-1', 'direction_t+1'], inplace=True)
df_1d['direction_t-1'] = df_1d['direction_t-1'].astype(int)
df_1d['direction_t+1'] = df_1d['direction_t+1'].astype(int)

# Saving the DataFrame as a parquet file

df_1m.to_parquet('../data/02-analysis_data/cleaned_data_1m.parquet', index=False)
df_1h.to_parquet('../data/02-analysis_data/cleaned_data_1h.parquet', index=False)
df_1d.to_parquet('../data/02-analysis_data/cleaned_data_1d.parquet', index=False)
