==> scripts/04-test_analysis_data.ipynb <==
#### Preamble ####
# Purpose: Validate all of the variables found the dataset
# Author: Jiazhou(Justin) Bi
# Date: 16 Nov 2024
# Contact: justin.bi@mail.utoronto.ca
# License: MIT
# Pre-requisites: see requirements.txt
# Any other information needed? None

# Loading the Dataset

import pandas as pd
df_1m = pd.read_parquet('../data/02-analysis_data/cleaned_data_1m.parquet')
df_1h = pd.read_parquet('../data/02-analysis_data/cleaned_data_1h.parquet')
df_1d = pd.read_parquet('../data/02-analysis_data/cleaned_data_1d.parquet')
df_1h.head()

# Open

The open price should always be lower than or equal to the high price, and always higher or equal to the low price. This subsection will check if this logic is correct.

if all((df_1m['open'] <= df_1m['high']) & (df_1m['open'] >= df_1m['low'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1h['open'] <= df_1h['high']) & (df_1h['open'] >= df_1h['low'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1d['open'] <= df_1d['high']) & (df_1d['open'] >= df_1d['low'])):
    print('PASS')
else:
    print('FAIL')

Originally, missing values were found here. Fixed them in 03-clean_data.ipynb, and they are fixed now.

# High

The high price should always be higher or equal to the open, low price, and close. This subsection will check if this logic is correct.

if all((df_1m['high'] >= df_1m['open']) & (df_1m['high'] >= df_1m['low']) &(df_1m['high'] >= df_1m['close'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1h['high'] >= df_1h['open']) & (df_1h['high'] >= df_1h['low']) &(df_1h['high'] >= df_1h['close'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1d['high'] >= df_1d['open']) & (df_1d['high'] >= df_1d['low']) &(df_1d['high'] >= df_1d['close'])):
    print('PASS')
else:
    print('FAIL')

# Low

The low price should always be lower or equal to the open, high price, and close. This subsection will check if this logic is correct.


if all((df_1m['low'] <= df_1m['open']) & (df_1m['low'] <= df_1m['high']) &(df_1m['low'] <= df_1m['close'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1h['low'] <= df_1h['open']) & (df_1h['low'] <= df_1h['high']) &(df_1h['low'] <= df_1h['close'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1d['low'] <= df_1d['open']) & (df_1d['low'] <= df_1d['high']) &(df_1d['low'] <= df_1d['close'])):
    print('PASS')
else:
    print('FAIL')

# Close

The close price should always be lower than or equal to the high price, and always higher or equal to the low price. Although this has been proven by the subsections above, we will double-check  the data. This subsection will check if this logic is correct.

if all((df_1m['close'] <= df_1m['high']) & (df_1m['close'] >= df_1m['low'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1h['close'] <= df_1h['high']) & (df_1h['close'] >= df_1h['low'])):
    print('PASS')
else:
    print('FAIL')

if all((df_1d['close'] <= df_1d['high']) & (df_1d['close'] >= df_1d['low'])):
    print('PASS')
else:
    print('FAIL')

# Volume

Volume can be any number that is non-negative. This subsection will verify this logic.

if all(df_1d['volume'] >= 0):
    print('PASS')
else:
    print('FAIL')

# Was_missing

This column is generated in 03-clean_data.ipynb to serve as a flag to indicate if the row exists originally from the raw dataset, or was missing and interpolated. It should either be 0 (originally exist) or 1(originally missing). This subsection will verify this. Note that there was no missing values for the 1-day dataset, so this column does not exist in that dataset.

unique_values = df_1m['was_missing'].unique()
print(unique_values)
print(df_1m['was_missing'].isna().sum())

unique_values = df_1h['was_missing'].unique()
print(unique_values)
print(df_1h['was_missing'].isna().sum())

# Direction

This column is generated in 03-clean_data.ipynb to serve as an indicator whether the closing price has increased, decreased, or remained the same for the previous timestamp. It should either be -1 (decreased), 0 (remianed the same), or 1(increased). This subsection will verify this.

unique_values = df_1m['direction_t-1'].unique()
print(unique_values)
print(df_1m['direction_t-1'].isna().sum())

unique_values = df_1m['direction_t+1'].unique()
print(unique_values)
df_1m['direction_t+1'] = df_1m['direction_t+1'].astype(int)
print(df_1m['direction_t+1'].isna().sum())

unique_values = df_1h['direction_t-1'].unique()
print(unique_values)
print(df_1h['direction_t-1'].isna().sum())

unique_values = df_1h['direction_t+1'].unique()
print(unique_values)
df_1h['direction_t+1'] = df_1h['direction_t+1'].astype(int)
print(df_1h['direction_t+1'].isna().sum())

unique_values = df_1d['direction_t-1'].unique()
print(unique_values)
print(df_1d['direction_t-1'].isna().sum())

unique_values = df_1d['direction_t+1'].unique()
print(unique_values)
df_1d['direction_t+1'] = df_1d['direction_t+1'].astype(int)
print(df_1d['direction_t+1'].isna().sum())
